+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## [0.7.2] - 2023-03
 ### New
 - Downloading: improve support for subtitles by making use of Representation.mimeType attributes
+- New crate feature `native-tls` (enabled by default) which is passed through to the `reqwest`
+  crate.
+- New crate feature `socks` (enabled by default) which is passed through to the `reqwest` crate. It
+  enables SOCKS5 proxy support for HTTP/HTTPS requests.
 - Add `@segmentProfiles` attributes to Representation and AdaptationSet nodes.
 - Add `@dependencyId` attribute to Representation nodes.
 - Add `@qualityRanking` attribute to Representation nodes.